client changes updated

+++ api/views.py
@@ -54,11 +54,11 @@ class ParkZoneViewSet(viewsets.ModelViewSet):
             return Response({"detail": "Invalid vehicle type"}, status=status.HTTP_400_BAD_REQUEST)
         return super().create(request, *args, **kwargs)
 
-    def update(self, request, *args, **kwargs):
-        vehicle_type = request.data.get('vehicle_type')
-        if vehicle_type not in ['bike', 'car', 'heavy']:
-            return Response({"detail": "Invalid vehicle type"}, status=status.HTTP_400_BAD_REQUEST)
-        return super().update(request, *args, **kwargs)
+    # def update(self, request, *args, **kwargs):
+    #     vehicle_type = request.data.get('vehicle_type')
+    #     if vehicle_type not in ['bike', 'car', 'heavy']:
+    #         return Response({"detail": "Invalid vehicle type"}, status=status.HTTP_400_BAD_REQUEST)
+    #     return super().update(request, *args, **kwargs)
 
 def create_ticket_code():
     return ''.join(random.choices(string.ascii_lowercase + string.digits, k=6))
@@ -70,36 +70,31 @@ class ReservationView(APIView):
     model = Reservation
     serializer_class = ReservationSerializer
 
-    def get(self, request):
+    def post(self, request, *args, **kwargs):
         try:
-            user_reservation = self.model.objects.filter(customer=request.user, checked_out=False).first()
-            if user_reservation:
-                serializer = self.serializer_class(user_reservation)
-                return Response(serializer.data)
-            else:
-                return Response({'message': 'No active reservation found'}, status=status.HTTP_404_NOT_FOUND)
+            existing_reservation = self.model.objects.get(customer=request.user, checked_out=False)
+            return Response({'message': 'You already have an active reservation'}, status=status.HTTP_400_BAD_REQUEST)
         except self.model.DoesNotExist:
-            return Response({'message': 'No active reservation found'}, status=status.HTTP_404_NOT_FOUND)
+            pass  # User does not have an active reservation, continue with reservation creation
 
-    def post(self, request):
-        serializer = self.serializer_class(data=request.data)
+        pk = kwargs.get("pk")
+        serializer = self.serializer_class(data=request.data, context={'pk': pk})
         if serializer.is_valid():
             start_time = serializer.validated_data['start_time']
             end_time = serializer.validated_data['finish_time']
-            parking_zone_id = serializer.validated_data['parking_zone']
-            vehicle_type = serializer.validated_data.get('vehicle_type')
+            parking_zone_id = pk
 
             if start_time >= end_time:
                 return Response({'message': 'End time must be after start time'}, status=status.HTTP_400_BAD_REQUEST)
 
-            # Check if there is an active reservation for the same vehicle type
-            active_reservations = self.model.objects.filter(customer=request.user, checked_out=False, parking_zone__vehicle_type=vehicle_type)
-            if active_reservations.exists():
-                return Response({'message': 'You already have an active reservation for this vehicle type'}, status=status.HTTP_400_BAD_REQUEST)
-
             parking_zone = get_object_or_404(ParkZone, id=parking_zone_id)
             if parking_zone.vacant_slots == 0:
                 return Response({'message': 'Parking Zone Full!'}, status=status.HTTP_400_BAD_REQUEST)
+
+            # Check if the parking zone is for the correct vehicle type
+            vehicle_type = serializer.validated_data.get('vehicle_type')
+            if vehicle_type != parking_zone.vehicle_type:
+                return Response({'message': 'Invalid vehicle type for this parking zone'}, status=status.HTTP_400_BAD_REQUEST)
 
             ticket_code = create_ticket_code()
             while self.model.objects.filter(ticket_code=ticket_code).exists():
@@ -119,39 +114,51 @@ class ReservationView(APIView):
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
 
-    def delete(self, request):
-        vehicle_type = request.data.get('vehicle_type')  # Get the selected vehicle type from request data
+class CancelReservationView(APIView):
+    authentication_classes = [BasicAuthentication, TokenAuthentication]
+    permission_classes = [permissions.IsAuthenticated]
+    model = Reservation
+
+    def delete(self, request, pk, *args, **kwargs):
         try:
-            user_reservation = self.model.objects.get(customer=request.user, checked_out=False, parking_zone__vehicle_type=vehicle_type)
+            user_reservation = self.model.objects.get(customer=request.user, pk=pk)
+            if user_reservation.checked_out:
+                return Response({'message': 'Reservation has already been checked out'}, status=status.HTTP_400_BAD_REQUEST)
+            parking_zone = user_reservation.parking_zone
             with transaction.atomic():
-                parking_zone = user_reservation.parking_zone
                 parking_zone.occupied_slots -= 1
                 parking_zone.vacant_slots += 1
                 parking_zone.save()
                 user_reservation.delete()
             return Response({'message': 'Reservation deleted successfully'}, status=status.HTTP_204_NO_CONTENT)
         except self.model.DoesNotExist:
-            return Response({'message': 'No active reservation found for the selected vehicle type'}, status=status.HTTP_404_NOT_FOUND)
+            return Response({'message': 'Reservation not found'}, status=status.HTTP_404_NOT_FOUND)
+
         
-class ParkingZoneSearchView(APIView):
-    authentication_classes = [BasicAuthentication, TokenAuthentication]
-    permission_classes = [permissions.IsAuthenticated]
+class ParkZoneSearchView(APIView):
+    def get(self, request):
+        state_id = request.query_params.get('state')
+        district_id = request.query_params.get('district')
+        location_id = request.query_params.get('location')
+        vehicle_type = request.query_params.get('vehicle_type')
 
-    def post(self, request):
-        location = request.data.get('location')
-        vehicle_type = request.data.get('vehicle_type')
+        if not all([state_id, district_id, location_id, vehicle_type]):
+            return Response({'error': 'Please provide all search parameters.'}, status=400)
 
-        if location and vehicle_type:
-            zones = ParkZone.objects.filter(location__icontains=location, vehicle_type__iexact=vehicle_type)
-        elif location:
-            zones = ParkZone.objects.filter(location__icontains=location)
-        elif vehicle_type:
-            zones = ParkZone.objects.filter(vehicle_type__iexact=vehicle_type)
-        else:
-            return Response({'error': 'Please provide location and/or vehicle_type parameters.'}, status=400)
+        try:
+            zones = ParkZone.objects.filter(
+                state_id=state_id,
+                district_id=district_id,
+                location_id=location_id,
+                vehicle_type=vehicle_type
+            )
+        except ParkZone.DoesNotExist:
+            return Response({'error': 'No matching ParkZone found.'}, status=404)
 
         serializer = ParkZoneSearchSerializer(zones, many=True)
         return Response(serializer.data)
+
+
     
 class TicketPdfView(APIView):
     authentication_classes = [BasicAuthentication, TokenAuthentication]
@@ -178,27 +185,20 @@ class CheckOutView(APIView):
     authentication_classes = [BasicAuthentication, TokenAuthentication]
     permission_classes = [permissions.IsAuthenticated]
 
-    def post(self, request):
-        vehicle_type = request.data.get('vehicle_type')
-        if vehicle_type not in ['bike', 'car', 'heavy']:
-            return Response({'message': 'Invalid vehicle type'}, status=status.HTTP_400_BAD_REQUEST)
-
+    def post(self, request, pk):
         try:
-            reservation = Reservation.objects.filter(customer=request.user, checked_out=False, vehicle_type=vehicle_type).first()
-            if reservation:
-                reservation.checked_out = True
-                reservation.save()
+            reservation = Reservation.objects.get(customer=request.user, checked_out=False, pk=pk)
+            reservation.checked_out = True
+            reservation.save()
 
-                parking_zone = reservation.parking_zone
-                parking_zone.occupied_slots -= 1
-                parking_zone.vacant_slots += 1
-                parking_zone.save()
+            parking_zone = reservation.parking_zone
+            parking_zone.occupied_slots -= 1
+            parking_zone.vacant_slots += 1
+            parking_zone.save()
 
-                return Response({'message': 'Successfully Checked Out'})
-            else:
-                return Response({'message': f'No Parking reservation exists for {request.user}'}, status=status.HTTP_404_NOT_FOUND)
-        except ObjectDoesNotExist:
-            return Response({'message': f'No Parking reservation exists for {request.user}'}, status=status.HTTP_404_NOT_FOUND)
+            return Response({'message': 'Successfully Checked Out'})
+        except Reservation.DoesNotExist:
+            return Response({'message': f'No Parking reservation exists for {request.user} with ID {pk}'}, status=status.HTTP_404_NOT_FOUND)
 
 class ChangePasswordView(APIView):
     permission_classes = [permissions.IsAuthenticated]
